refactor: index files and create folders
