+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "myapp",
+  "name": "lira",
   "version": "0.0.1",
   "private": true,
   "main": "expo-router/entry",
