+++ src/app/(app)/settings.tsx
@@ -1,18 +1,15 @@
-/* eslint-disable react/react-in-jsx-scope */
 import { Env } from '@env';
 import { useColorScheme } from 'nativewind';
+import React from 'react';
 
-import { Item } from '@/components/settings/item';
-import { ItemsContainer } from '@/components/settings/items-container';
-import { LanguageItem } from '@/components/settings/language-item';
-import { ThemeItem } from '@/components/settings/theme-item';
 import {
-  colors,
-  FocusAwareStatusBar,
-  ScrollView,
-  Text,
-  View,
-} from '@/components/ui';
+  Item,
+  ItemsContainer,
+  LanguageItem,
+  ThemeItem,
+} from '@/components/features';
+import { FocusAwareStatusBar, ScrollView, Text, View } from '@/components/ui';
+import colors from '@/components/ui/colors';
 import { Github, Rate, Share, Support, Website } from '@/components/ui/icons';
 import { translate, useAuth } from '@/lib';
