+++ src/app/[...messing].tsx
@@ -1,4 +1,5 @@
 import { Link, Stack } from 'expo-router';
+import React from 'react';
 
 import { Text, View } from '@/components/ui';
