+++ src/app/login.tsx
@@ -1,8 +1,7 @@
 import { useRouter } from 'expo-router';
 import React from 'react';
 
-import type { LoginFormProps } from '@/components/login-form';
-import { LoginForm } from '@/components/login-form';
+import { LoginForm, type LoginFormProps } from '@/components/features';
 import { FocusAwareStatusBar } from '@/components/ui';
 import { useAuth } from '@/lib';
