+++ src/components/features/index.ts
@@ -1,0 +1,9 @@
+export * from './login/login-form';
+export * from './settings/item';
+export * from './settings/items-container';
+export * from './settings/language-item';
+export * from './settings/theme-item';
+export * from './style/buttons';
+export * from './style/colors';
+export * from './style/inputs';
+export * from './style/typography';
