+++ src/components/features/login/login-form.test.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 
 import { cleanup, screen, setup, waitFor } from '@/lib/test-utils';
 
-import type { LoginFormProps } from './login-form';
-import { LoginForm } from './login-form';
+import { LoginForm, type LoginFormProps } from './login-form';
 
 afterEach(cleanup);
