+++ src/components/features/settings/language-item.tsx
@@ -2,8 +2,7 @@ import * as React from 'react';
 
 import type { OptionType } from '@/components/ui';
 import { Options, useModal } from '@/components/ui';
-import { useSelectedLanguage } from '@/lib';
-import { translate } from '@/lib';
+import { translate, useSelectedLanguage } from '@/lib';
 import type { Language } from '@/lib/i18n/resources';
 
 import { Item } from './item';
