+++ src/components/ui/primitives/button/button.test.tsx
@@ -4,7 +4,7 @@ import { Text } from 'react-native';
 
 import { cleanup, render, screen, setup } from '@/lib/test-utils';
 
-import { Button } from './button';
+import { Button } from '../../primitives/button/button';
 
 afterEach(cleanup);
