+++ src/components/ui/primitives/focus-aware-status-bar.tsx
