+++ src/components/ui/primitives/image/image.tsx
