+++ src/components/ui/primitives/input/input.tsx
@@ -15,8 +15,8 @@ import {
 } from 'react-native';
 import { tv } from 'tailwind-variants';
 
-import colors from './colors';
-import { Text } from './text';
+import colors from '../../colors';
+import { Text } from '../text';
 
 const inputTv = tv({
   slots: {
