+++ src/lib/index.tsx
@@ -1,4 +1,5 @@
 export * from './auth';
 export * from './hooks';
 export * from './i18n';
+export * from './use-theme-config';
 export * from './utils';
